refactor variant comparison tests
